tfidf/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via cryptography
chardet==5.2.0
    # via -r requirements.in
charset-normalizer==3.4.2
    # via requests
cryptography==45.0.3
    # via social-auth-core
dawg-python==0.7.2
    # via pymorphy2
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==5.2.1
    # via
    #   -r requirements.in
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   social-auth-app-django
djangorestframework==3.16.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-simplejwt==5.5.0
    # via
    #   -r requirements.in
    #   djoser
djoser==2.3.1
    # via -r requirements.in
docopt==0.6.2
    # via pymorphy2
drf-spectacular==0.28.0
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.24.0
    # via drf-spectacular
jsonschema-specifications==2025.4.1
    # via jsonschema
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==25.0
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.9.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pymorphy2==0.9.1
    # via -r requirements.in
pymorphy2-dicts-ru==2.4.417127.4579844
    # via pymorphy2
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.2
    # via drf-spectacular
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
social-auth-app-django==5.4.3
    # via djoser
social-auth-core==4.6.1
    # via social-auth-app-django
sqlparse==0.5.3
    # via django
typing-extensions==4.13.2
    # via
    #   asgiref
    #   referencing
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.4.0
    # via requests
